renaming HomePage in DiscoverPage
